frontend: server-rendered per-paper metadata on /abs/[...id]
The reader was a "use client" page, so every paper link unfurled with the
generic site card and search engines indexed one title for all papers. Split
the route: page.tsx is now a server component whose generateMetadata fetches
the paper's title/abstract from the API (3 s timeout, 10 min revalidate,
fail-safe to a generic "arXiv:<id>" title on 404/error) and renders the
unchanged client reader from PaperPage.tsx, passing the same params promise.
normalizeArxivId moves to lib/arxiv-id.ts so both halves share it; API_BASE
is exported from lib/api.ts for the server-side fetch.
