+++ frontend/lib/api.ts
@@ -11,8 +11,9 @@ import type { Paper, PaperStatus, PaperSummary, ProcessingStatus, Section } from
 const USE_MOCK = process.env.NEXT_PUBLIC_USE_MOCK === "true";
 
 // Backend API base URL — env var wins; production builds fall back to the
-// Azure backend, dev builds to localhost.
-const API_BASE =
+// Azure backend, dev builds to localhost. Exported for the server-side
+// metadata fetch (lib/paper-metadata.ts), which must hit the same backend.
+export const API_BASE =
   process.env.NEXT_PUBLIC_API_URL ||
   (process.env.NODE_ENV === "production"
     ? "https://arxivisual-api.purplepond-ac9e2dc5.eastus2.azurecontainerapps.io"
